apps/image/requirements.txt
customtkinter
numpy
opencv-python
OpenEXR
piexif
Pillow
psd-tools
rembg
scipy
vtracer
